chore(config): FE-00 Split the different snapshots to facilitate review
This way it is easier to spot which ones fail when eslint fails.
